Day 22, only part 1 is working so far.
I made the critical mistake of not looking at my real input,
and implemented a solution that works perfectly for the test input,
but depends on my real input having the same layout as the test layout.
When I actually ran my program on my real input, I got bounds errors
because the real input isn't square, and the adjacency
map is all wrong too.

The easiest way to solve it may be to injest the faces, manually fold and
rotate them to match the test input's orientations, then proceed.